requirements.txt
async-timeout==4.0.2
certifi==2023.5.7
charset-normalizer==3.1.0
decorator==4.4.2
idna==3.4
imageio==2.31.1
imageio-ffmpeg==0.4.8
lazy_loader==0.2
moviepy==1.0.3
networkx==3.1
numpy==1.24.3
opencv-python==4.7.0.72
packaging==23.1
Pillow==9.5.0
proglog==0.1.10
PyWavelets==1.4.1
redis==4.5.5
requests==2.31.0
scikit-image==0.21.0
scipy==1.10.1
tifffile==2023.4.12
tqdm==4.65.0
urllib3==2.0.3
